:sparkles: Updates how server sets the locale
Before coding the switcher itself, I must allow Next to select the locale based
on a manual selected item.

In this commit, I've added all required logic to allow NextJS parsging cookies
and selecting the required locale based on this.

Now, we can change the site language based on the locale cookie.